Refactor callback
refactor: extracted the open_video_import_dialog because it is used in more than one python file
